rearrange: create ls module
